requirements.txt
asgiref==3.2.3
Django==3.0
django-extensions==2.2.5
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.9
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
